Distributed cilium

+++ microk8s-resources/actions/enable.cilium.sh
@@ -10,16 +10,38 @@ if ! [ "${ARCH}" = "amd64" ]; then
   exit 1
 fi
 
-"$SNAP/microk8s-enable.wrapper" helm
+read -ra CILIUM_VERSION <<< "$1"
+if [ -z "$CILIUM_VERSION" ]; then
+  CILIUM_VERSION="v1.6"
+fi
+CILIUM_ERSION=$(echo $CILIUM_VERSION | sed 's/v//g')
 
-echo "Restarting kube-apiserver"
-refresh_opt_in_config "allow-privileged" "true" kube-apiserver
-sudo systemctl restart snap.${SNAP_NAME}.daemon-apiserver
+if [ -f "${SNAP_DATA}/bin/cilium-$CILIUM_ERSION" ]
+then
+  echo "Cilium version $CILIUM_VERSION is already installed."
+  exit 1
+fi
 
-# Reconfigure kubelet/containerd to pick up the new CNI config and binary.
-echo "Restarting kubelet"
-refresh_opt_in_config "cni-bin-dir" "\${SNAP_DATA}/opt/cni/bin/" kubelet
-sudo systemctl restart snap.${SNAP_NAME}.daemon-kubelet
+echo "Enabling Cilium"
+CILIUM_DIR="cilium-$CILIUM_ERSION"
+SOURCE_URI="https://github.com/cilium/cilium/archive"
+CILIUM_CNI_CONF="plugins/cilium-cni/05-cilium-cni.conf"
+CILIUM_LABELS="k8s-app=cilium"
+NAMESPACE=kube-system
+
+echo "Fetching cilium version $CILIUM_VERSION."
+sudo mkdir -p "${SNAP_DATA}/tmp/cilium"
+(cd "${SNAP_DATA}/tmp/cilium"
+sudo "${SNAP}/usr/bin/curl" -L $SOURCE_URI/$CILIUM_VERSION.tar.gz -o "$SNAP_DATA/tmp/cilium/cilium.tar.gz"
+if ! sudo gzip -f -d "$SNAP_DATA/tmp/cilium/cilium.tar.gz"; then
+  echo "Invalid version \"$CILIUM_VERSION\". Must be a branch on https://github.com/cilium/cilium."
+  exit 1
+fi
+
+sudo tar -xf "$SNAP_DATA/tmp/cilium/cilium.tar" "$CILIUM_DIR/install" "$CILIUM_DIR/$CILIUM_CNI_CONF")
+sudo mv "$SNAP_DATA/args/cni-network/cni.conf" "$SNAP_DATA/args/cni-network/10-kubenet.conf" 2>/dev/null || true
+sudo mv "$SNAP_DATA/args/cni-network/flannel.conflist" "$SNAP_DATA/args/cni-network/20-flanneld.conflist" 2>/dev/null || true
+sudo cp "$SNAP_DATA/tmp/cilium/$CILIUM_DIR/$CILIUM_CNI_CONF" "$SNAP_DATA/args/cni-network/05-cilium-cni.conf"
 
 set_service_not_expected_to_start flanneld
 sudo systemctl stop snap.${SNAP_NAME}.daemon-flanneld
@@ -31,39 +53,11 @@ if grep -qE "bin_dir.*SNAP}\/" $SNAP_DATA/args/containerd-template.toml; then
   sudo systemctl restart snap.${SNAP_NAME}.daemon-containerd
 fi
 
-echo "Enabling Cilium"
-
-read -ra CILIUM_VERSION <<< "$1"
-if [ -z "$CILIUM_VERSION" ]; then
-  CILIUM_VERSION="v1.6"
-fi
-CILIUM_ERSION=$(echo $CILIUM_VERSION | sed 's/v//g')
-
-if [ -f "${SNAP_DATA}/bin/cilium-$CILIUM_ERSION" ]
+if ! [ -e "$SNAP_DATA/var/lock/clustered.lock" ]
 then
-  echo "Cilium version $CILIUM_VERSION is already installed."
-else
-  CILIUM_DIR="cilium-$CILIUM_ERSION"
-  SOURCE_URI="https://github.com/cilium/cilium/archive"
-  CILIUM_CNI_CONF="plugins/cilium-cni/05-cilium-cni.conf"
-  CILIUM_LABELS="k8s-app=cilium"
-  NAMESPACE=kube-system
-
-  echo "Fetching cilium version $CILIUM_VERSION."
-  sudo mkdir -p "${SNAP_DATA}/tmp/cilium"
-  (cd "${SNAP_DATA}/tmp/cilium"
-  sudo "${SNAP}/usr/bin/curl" -L $SOURCE_URI/$CILIUM_VERSION.tar.gz -o "$SNAP_DATA/tmp/cilium/cilium.tar.gz"
-  if ! sudo gzip -f -d "$SNAP_DATA/tmp/cilium/cilium.tar.gz"; then
-    echo "Invalid version \"$CILIUM_VERSION\". Must be a branch on https://github.com/cilium/cilium."
-    exit 1
-  fi
-  sudo tar -xf "$SNAP_DATA/tmp/cilium/cilium.tar" "$CILIUM_DIR/install" "$CILIUM_DIR/$CILIUM_CNI_CONF")
-
-  sudo mv "$SNAP_DATA/args/cni-network/cni.conf" "$SNAP_DATA/args/cni-network/10-kubenet.conf" 2>/dev/null || true
-  sudo mv "$SNAP_DATA/args/cni-network/flannel.conflist" "$SNAP_DATA/args/cni-network/20-flanneld.conflist" 2>/dev/null || true
-  sudo cp "$SNAP_DATA/tmp/cilium/$CILIUM_DIR/$CILIUM_CNI_CONF" "$SNAP_DATA/args/cni-network/05-cilium-cni.conf"
-
-  # Generate the YAMLs for Cilium and apply them
+  # Only on the master
+  "$SNAP/microk8s-enable.wrapper" helm
+    # Generate the YAMLs for Cilium and apply them
   (cd "${SNAP_DATA}/tmp/cilium/$CILIUM_DIR/install/kubernetes"
   ${SNAP_DATA}/bin/helm template cilium \
       --namespace $NAMESPACE \
@@ -77,6 +71,16 @@ else
   sudo mkdir -p "$SNAP_DATA/actions/cilium/"
   sudo cp "$SNAP_DATA/tmp/cilium/$CILIUM_DIR/install/kubernetes/cilium.yaml" "$SNAP_DATA/actions/cilium.yaml"
   sudo sed -i 's;path: \(/var/run/cilium\);path: '"$SNAP_DATA"'\1;g' "$SNAP_DATA/actions/cilium.yaml"
+
+  echo "Restarting kube-apiserver"
+  refresh_opt_in_config "allow-privileged" "true" kube-apiserver
+  restart_service apiserver
+
+  # Reconfigure kubelet/containerd to pick up the new CNI config and binary.
+  echo "Restarting kubelet"
+  refresh_opt_in_config "cni-bin-dir" "\${SNAP_DATA}/opt/cni/bin/" kubelet
+  restart_service kubelet
+
 
   ${SNAP}/microk8s-status.wrapper --wait-ready >/dev/null
   echo "Deploying $SNAP_DATA/actions/cilium.yaml. This may take several minutes."
@@ -92,8 +96,10 @@ else
   sudo chmod +x "$SNAP_DATA/bin/"
   sudo chmod +x "$SNAP_DATA/bin/cilium-$CILIUM_ERSION"
   sudo ln -s $SNAP_DATA/bin/cilium-$CILIUM_ERSION $SNAP_DATA/bin/cilium
-
-  sudo rm -rf "$SNAP_DATA/tmp/cilium"
 fi
+
+nodes_addon cilium enable
+
+sudo rm -rf "$SNAP_DATA/tmp/cilium"
 
 echo "Cilium is enabled"
